// src/app/models/superhero-creator.model.ts
export type LocalHeroInput = {
  name: string;
  imageUrl?: string;
  intelligence?: string;
  strength?: string;
  speed?: string;
  durability?: string;
  power?: string;
  combat?: string;
  fullName?: string;
  alterEgos?: string;
  aliasesCsv?: string;
  placeOfBirth?: string;
  firstAppearance?: string;
  publisher?: string;
  alignment?: string;
  gender?: string;
  race?: string;
  height0?: string;
  height1?: string;
  weight0?: string;
  weight1?: string;
  eyeColor?: string;
  hairColor?: string;
  occupation?: string;
  base?: string;
  groupAffiliation?: string;
  relatives?: string;
};